Updates readme. Renames component file.
