+++ packages/ia-icons/index.js
@@ -1,3 +1,3 @@
-import IAIcon from './src/ia-icons';
+import IAIcon from './src/ia-icon';
 
 export default IAIcon;
