+++ packages/ia-icons/src/ia-icon.js
